Update patch set 2
Patch Set 2:

(17 comments)

Much of the generality here is needed as part of the pprof command but is not needed when the runtime is generating its own profiles. For example we know perfectly well what the byte order and word size are. I think it will help a lot to cut this down to the job at hand. As cmd/pprof/README warns, most of pprof is arguably overgeneralized and should not be taken as an example to follow in other packages, including this one.

Patch-set: 2