Update patch set 8
Patch Set 8: Code-Review+1

Are you happy with the changes made to the generic parts Russ? If so I'm happy to +2 this.

I'd like to clean up some of the comments in p256_s390x.go, but they are more or less the same as in p256_amd64.go and so I think I'll attempt both in a follow up CL.

Patch-set: 8