Update patch set 8
Patch Set 8: Code-Review+2

> This is fine for now, but in Go 1.9 let's refactor the amd64
 > support out to its own file so p256.go is the generic file.

If I understand correctly this is already the case. There is a specialized pure-Go implementation of P256 in p256.go that is not built on amd64 because the amd64 assembly doesn't require a feature check.

There is also a general implementation that supports arbitrary curves in elliptic.go which is used by the tests as an implementation to check the specialized P256 implementations against.

Patch-set: 8
Label: Code-Review=+2