Update patch set 1
Patch Set 1:

> Wait, if yours is inlined, I'm not sure why it's slower actually.

It seems that the compiler is not smart enough to realize that it doesn't need to emit `m = -1` and then check that it was < 0. Instead it should skip that part. I'll file a issue about this.

My solution has 4 extra instructions over your version, which causes mine to be 5% slower.

Patch-set: 1