cmd/link: always use relro for external linking on ELF
Increasingly the default toolchain installed on ELF-based systems is
defaulting to PIE.  When we don't use relro mode for a PIE, the
resulting program gets a DT_TEXTREL dynamic tag, which is inefficient
and reportedly does not work at all on some systems.  Using relro mode
avoids the problem.

Making this change revealed that the Go objdump tool does not know how
to find the pclntab for an executable built with relro, so fix that too.

Update #17847.

Change-Id: Ibc1a12295de3c061b26613adb86b6e10f3f0f9a8