Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Ooohh.. I can just imagine the test failures that are going to occur because we changed the spelling of "unmarshalled". Too many tests do: 
>  strings.Contains(err.Error(), "some string")
> 
> I'm all for consistency. Go ahead break these bad tests.

That would be unnecessary churn for our users.
I think we should preserve the existing error message as much as possible.

If the old spelling is a typo, then I agree we should fix the error message,
but if it's also a valid spelling, I don't think we should change it in
error messages for the sake of consistency.

Patch-set: 2