Update patch set 10
Patch Set 10:

> What's the status of this? It's still marked DO NOT SUBMIT.

As Quentin noted, one drawback with this approach is there is a race condition if the signal arrives just before the i/o syscall has started or just before it has completed. However, I don't know how we could fix this issue. I recall Russ once suggested that we could use special code in note handler that changes the PC if it is in a certain range.

That said, the net tests (with timeout tests enabled) and net/http tests are passing successfully. Sometimes, when cycling through the tests, I get a "write /net/tcp/clone: connection refused" error, which might be due to the race condition or might be unrelated.

Patch-set: 10
Label: Code-Review=0