Update patch set 11
Patch Set 11:

I'd tend to suggest that we submit this experimental change as is, as it would likely bring the three Plan 9 builders back to "ok", which is already a step forward. It would also allow us to enable a lot of tests. Then we could work on fixing the remaining issues.

Patch-set: 11
Reviewer: Gerrit User 5060 <5060@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0