Update patch set 2
Patch Set 2:

The old version uses MULSD and the new version uses DIVSD. I do not think its possible for the compiler to optimize divs to muls in general for floating point arithmetic with constants without loss of precision and with a speedup. For some divisions by powers of 2 it might be possible.

Patch-set: 2