Update patch set 1
Patch Set 1:

Oh, so is the reason this works on elf is because of the way we force all relocations against functions to be against a local symbol? I don't know if replicating that (or using any equivalent of -Bsymbolic-functions that the macho linker has) would be better or worse than this.

Should machoreloc1 explode if Xadd != 0? Seems better than just ignoring it...

Patch-set: 1