[dev.garbage] runtime: trace object logic
DO NOT REVIEW
For debugging it is useful trace how an object is
reached. This in turn is helpful in understanding why
an object is or is not marked as public. These routines
monitor the write barrier and whenever an object being
monitored is about to be placed in a slot the object
holding the slot is then monitored.

Change-Id: Ie8c3ccd5dbb4c516a756dfd656aff5332598705b