[dev.garbage] runtime: add ROC recycle calls
There are several places where a G can potentially lose a
P. Parking, syscalls, and scheduler calls are such places.
Currently we simply abort the ROC epoch by setting the
span's startindex to freeindex. This makes all allocated
object public even if they have not been made public and
are no longer reachable from the current goroutine's stack.
This CL scans the goroutines stack making any reachable
objects public and then recycles the remaining unreachable
local objects. At the moment syscall does not take advantage
of this optimization.

As a check against bugs we also ensure that if a span
was used for allocation in this epoch it was used by the
current goroutine.

Change-Id: I2cdc6770b5e2794271bae96690569e950fcb2c46