Update patch set 1
Patch Set 1:

An alternative fix for the original issue is to
make cmd/go detect glibc and give out a warning
if static linking is requested.

But that's probably a little hard to do. How could
we detect glibc reliably?

PS: I do think we should address some of the linker
warnings.

Patch-set: 1