[dev.garbage] runtime: hacks DO NOT REVIEW
While debugging on the 1.7 base we encountered some
problems that are better fixed at tip. This CL works
around some of the problems.

We also comment out TestArrayHash test in hash_test.go
which is a real bug that needs to be investigated but
we can also do that at tip.

GODEBUG=gcroc=n sets debug.gcroc to an int32.

Change-Id: I4374dce4e0e06c93d5c5350f547dc9e8c131906a