Update patch set 3
Patch Set 3:

I think the first step has to be writing down rules for how symlinks should work in GOPATH, and get everyone to agree on those rules.  Right now we have a very simple rule that anyone can understand: don't put a symlink in GOPATH.  If we want to change the rule in any way, I think we need to write down what the new rule should be.

Patch-set: 3