Update patch set 1
Patch Set 1:

I thought about it, but with all of the stmt refs that get strewn around and closed in a loop, the only way I'd feel comfortable ensure that is to wrap each driver statement in something like:
sturct {
  driver.Stmt
  closed bool
}
And then to check closed each time before calling Close().
...

Actually, looking around a bit more, you have this comment:
		// TODO(bradfitz): after Go 1.2, closing driver.Stmts
		// should be moved to driverStmt, using unique
		// *driverStmts everywhere (including from
		// *Stmt.connStmt, instead of returning a
		// driver.Stmt), using driverStmt as a pointer
		// everywhere, and making it a finalCloser.

If we did this we could indeed check this for the driver. Would this be in scope for the freeze?

Patch-set: 1