Update patch set 1
Patch Set 1: Code-Review+1

LGTM

The key here is that the Default Converter first checks for IsValue, and then uses reflect to determine if there is indirect pointers or variations of a type. I think the least surprising thing is for direct types string, int, to also work for custom types for all valid types, not just some of the valid default types is is the case now.

No drivers should be impacted as we are not passing any new types to the drivers. Nor will it be significantly more expensive because the reflect checks are already being done for some of the types. sql just works with more typed data for the user.

Patch-set: 1
Reviewer: Gerrit User 5080 <5080@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1