Update patch set 2
Patch Set 2:

I support this change, but Russ was worried about the quadratic behavior of a similar change:

https://golang.org/cl/24469

The O(N^2) behavior usually isn't a problem because N is small, but it is possible to construct programs that get stuck on it, so I have a hard time arguing against the complaint.

I've been slow to respond to this CL because I have been looking for a small (lines of code) data structure to replace the map that removes the quadratic behavior. Theoretically it's possible: the old data structure that is being replaced is still good for reading, so what's needed is some kind of append-only map.

A map + a fixed overflow array amortizes the cost but doesn't totally fix it.

Languages that prioritize immutable data structures (haskell, scala) achieve O(log N) map inserts without modifying the original data using a Hash array mapped trie:

https://infoscience.epfl.ch/record/64398/files/idealhashtrees.pdf

...but I'm not too keen on dragging an implementation of that into reflect (or the runtime package, which could do this work for reflect if we wanted), unless there were some other compelling uses of it.

Patch-set: 2