Update patch set 1
Patch Set 1:

(3 comments)

> Patch Set 1: Code-Review+2
> 
> (3 comments)
> 
> Thanks.
> 
> I note that this doesn't handle the case of a pointer to a pointer.  What do think of emitting a weak relocation if t.ptrTo != nil?  In other words, if the type has already been created for some other reason.

Done. This catches some unusual types in cmd/compile like ***ssa.Block and **[776]gc.ProgInfo.

Note that it's not a complete mapping for ptrToThis, because it requires the compiler to know about the type, and it's possible some other compilation unit knows about a ptrTo type we don't.

Patch-set: 1
Label: Code-Review=0