Update patch set 1
Patch Set 1:

I replied on the issue tracker (as below). For that reason, trace=all may not work either...

I tried to trace n64_sendfile on linux-mips64 gomote machine. Now it sees the syscall,

=== RUN   TestLinuxSendfile
--- FAIL: TestLinuxSendfile (0.22s)
	fs_test.go:1122: no sendfile system call found in:
		[pid  4003] n64_sendfile()              = 22
		[pid  4006] +++ exited with 0 +++
		[pid  4005] +++ exited with 0 +++
		[pid  4004] +++ exited with 0 +++
		+++ exited with 0 +++
but it fails to get the arguments as it uses different ABI, so the regexp match later in the test function fails.

Also changing to trace n64_sendfile would break systems on which strace is using the 64-nit ABI (because the syscall is named sendfile there...). (One possible way is to try both...)

Patch-set: 1
Label: Code-Review=0