cmd/compile: used shared epilogue
Instead of generating the epilogue at each return site (called "split"
epilogue in this CL description), generate it once and jump to it for
all other uses ("common" epilogue).

The split epilogue is better if the epilogue is short.  The common
epilogue is better when the epilogue is long.  Now that frame pointers
are enabled, we've pushed up the epilogue size so that it's probably
better to use a common one.

Saves 1.7% in binary size for cmd/go.

TODO: performance numbers
TODO: vary for amd64 vs. other archs?
TODO: split epilogue for frameless leaf functions?

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: Ic2446e2621cb36e9a2dee6635b28bfa0df6ca769