Update patch set 1
Patch Set 1:

> Martin, do you have plans for that?

tl;dr i do not think it will be possible for the non-ASCII case to ever be faster without generating binary bloat by e.g. inlining runtime.decoderune in every string iteration.


We possibly can get the runtime.decoderune to be a bit faster with some compiler optimizations that generate better code/branching.

However due to at least some function call overhead with the non inlined calls to runtime.decoderune for each loop iteration in the non-ASCII case here i do not see how we ever can be faster than the manually inlined version of the same code. We could always inline the non-ascii rune decoding but i would guess that is not a good trade off vs binary size.

We could let the compiler recognize special for loops like rune counting and generate explicit code for that but it feels to be to seldom and not to be such a major speedup that we could not just write the code explicitly for the places it matters.

Patch-set: 1