Update patch set 1
Patch Set 1:

I'm not too keen on a change that seems to just alter performance characteristics and not really improve it. One could argue it simplifies code, but it also makes the code for the byte and string versions diverge.

Is there any chance the non-ASCII performance could be improved as well, if only to 0?

Patch-set: 1
Label: Code-Review=0