Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Let me put it another way:
> 
> https://build.golang.org/ is happy without this change, and all.bash is happy without this change.

I don't want the tree peppered with silly -l=4 tests when the goal is to make that the default anyway, at which point the entire tree becomes a test suite for -l=4. If we want this to show up on the dashboard before this becomes the default, we should have a -l=4 builder just like we have a noopt builder.

Patch-set: 1