Update patch set 9
Patch Set 9:

Ryan, we are thinking about a different optimization, assuming that the certs in kSecTrustSettingsDomainSystem are never untrusted. If we see the same cert marked as untrusted in kSecTrustSettingsDomainAdmin or kSecTrustSettingsDomainUser, we'll drop it from our list, but we won't go looking for them explicitly. If there are 150 in system, 15 in admin, and 15 in user, which is roughly the case on my laptop, then this reduces (150+15+15)*3 lookups to (15+15)*2 lookups, which is a 10X reduction in overhead.

Before this CL at all, a simple Go program that fetched https://rsc.io/pdf took 0.2 seconds. With the first draft of this CL, it took 0.5 seconds. With the optimization I just described, it's back down to 0.2 seconds. 

This does assume that kSecTrustSettingsDomainSystem won't hold any certs marked as untrusted _there_, but that seems to be the case: when you untrust something in System Roots in Keychain Access, it gets copied into the System (or login) keychain and that copy is what gets marked untrusted.

Thoughts?

Patch-set: 9
Label: Code-Review=0