Update patch set 10
Patch Set 10:

> Ryan, we are thinking about a different optimization, assuming that
 > the certs in kSecTrustSettingsDomainSystem are never untrusted.

While not guaranteed by API contract or by Apple, that is in practice how it's worked out.

Apple has included roots with a 'kSecTrustSettingsTrustRoot' but then had explicit settings regarding trust. Examples of this are CNNIC, which is trusted in System, and which even in *distrusted* by *Admin or by *User, has a whitelist of certificates that overrides any trust settings.

WoSign has a similar whitelist. Apple just announced yesterday that they were taking action against StartCom, but based on expiration date, so I don't know how they plan to implement that.

So today that holds, because *System trust settings are stored in the Security.framework plist which is supposed to be read-only and treated as a protected system image, but there's no guarantee that a future Apple security update won't change that.

 > This does assume that kSecTrustSettingsDomainSystem won't hold any
 > certs marked as untrusted _there_, but that seems to be the case:
 > when you untrust something in System Roots in Keychain Access, it
 > gets copied into the System (or login) keychain and that copy is
 > what gets marked untrusted.

Correct, no _user_ modifications should be propagated, AIUI, certainly not via Keychain, but I'm not familiar enough with the low-level details to know if securityd (security server) has a path, since it's an elevated process, to modify k*System if called via API or via the 'security' tool.

Patch-set: 10