Update patch set 1
Patch Set 1:

> This will add a string (at best one byte, at worst the length of the string + 1 byte or so) to each sym, whether needed or not. This surely will blow up export data and possibly impact import and thus compile speed.

Thanks for the feedback.  I hadn't considered this.

> This needs to be done more carefully:
> 1) We should only write this information when needed (if we can tell from the context we don't need special bits in most cases)

What context can we use to determine whether to read a string for the linkname?

> 2) Either way, this needs a format version change so we are compatible with old files

This only changes the compiler-dependent export data.  Wont the compiler abort if the compiler version (commit hash) stored in the object doesn't exactly match its own version string?  If so, in what case is backwards compatibility for the compiler-dependent export data useful?

> 3) We need tests, incl. measurements showing impact (or no impact) on compilation performance and export data size.

Agreed.

Patch-set: 1