cmd/compile: allow *int to uintptr, uint to uint64 in global var declarations
defaultlit on its own skips returning OCONV when converting
from *int to uintptr, when we are in global scope, ie not in
a function. The problem also manifests in uint to uint64
conversion.

Exhibits:
package a

var (
  a uint
  b = a
  c = (uintptr)(b)
  d = (uint64)(b)
)

func f1() {
  var (
    a uint
    b = a
    c = (uintptr)(b)
    d = (uint64)(b)
  )
}

go tool compile a.go && echo "success"

* Before:
a.go:6: cannot use a (type uint) as type uint64 in assignment
a.go:7: cannot use a (type uint) as type uintptr in assignment

* After:
success

Fixes #13263

Change-Id: I5c6239e705f35a29c72e06042b26f6174b17e24b