Update patch set 1
Patch Set 1:

> Is there any chance the non-ASCII performance could be improved as
 > well, if only to 0?

I try to see if we further can improve performance of runtime.decoderune in 1.9 and then will ping this CL again so we can reevaluate the performance tradeoff. At least If we switch to asm i think it is possible to reach 0 difference.

However that would only make it relatively faster than the current version not faster than just coping the new code into RuneCountInString to avoid function call overhead.

Patch-set: 1