cmd/compile: randomize value order in block for testing
A little bit of compiler stress testing. Randomize the order
of the values in a block before every phase. This randomization
makes sure that we're not implicitly depending on that order.

Mostly passes all.bash. Not sure if this is worth checking in or not.
It would need some work (trigger on env variable, run on check builder?)
before it is ready to go in.

One failure is test/nilptr3.go. If we have:
  _ = *p
  _ = *p
then which nil ptr check gets eliminated is random.

Update #18169

DO NOT REVIEW
DO NOT SUBMIT

Change-Id: Idae792a23264bd9a3507db6ba49b6d591a608e83