[dev.garbage] runtime: add spanclass logic to ROC
Previous CLs added the ability to segregate objects
with pointers from objects without pointers into
different spans. The method spanclass.noscan() is
an efficient way to determine whether the object
lacks pointers and does not need to be scanned.
This change replaces the calls to hasPointers with
calls to noscan.

Change-Id: I7e7aad66c66478170b7d49b4029967ed3139ec53