Update patch set 1
Patch Set 1:

> For instance, certain libraries (CryptoPP e.g.) are restraining the
 > modulus and subgroup sizes to the ones specified in FIPS 186-4. It
 > may be a solution to avoid arbitrary length private keys.
 > Yet I'm not sure whether we should add such restrictions or not.

Nor I. But since I think I'm going to land this for 1.8, I don't think it would be viable in this change since it's such a late addition to the cycle and an API change.

 > Otherwise, you may want to add a comment to mention that the
 > attempts' limit is arbitrarily set to 10, maybe?

I think for any non-degenerate and non-tiny DSA group, 10 attempts brings the failure probability down to the point where CPUs probably take the wrong branch more often. As such, I wouldn't burden people with thinking about it. (Unless I'm getting my calculations wrong, in which case increasing 10 to a value that makes it essentially impossible is probably the correct answer. Please tell me if you think that the failure probability is meaningfully non-zero though.)

Patch-set: 1
Label: Code-Review=0