Update patch set 1
Patch Set 1:

This is the kind of thing that seems trivial but can have weird consequences. I'd be happy to do this at the beginning of a release cycle, but do we have evidence that there's both a need for this and that it's safe to do this late in the freeze?

Patch-set: 1