Update patch set 3
Patch Set 3:

This test is disabled here mainly for mips64, which already uses _cgo_try_pthread_create; it's currently skipped on mips anyway because cgo is not yet enabled for it.
The same machine is used for mips and mips64 buildbots (besides, they usually run at the same time) and it's weaker than mipsle and mips64le buildbots.
If attempts are 44 or less, the test passes, if more than 44, it timeouts.
(So all this is about letting all.bash pass on the current mips/mips64 buildbot machine.)

Patch-set: 3
Reviewer: Gerrit User 15150 <15150@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=0